feat: escaping
If drone is in obstacle when planning a route, the EGOPLANNER will be unable to plan.
Now, this new version of "ego_replan_fsm" will detect the position of drone when it receives a navigating target, and try to escape the obstacle area if it's trapped in it.
After escaping the area, it will continue to plan a route toward the original target.